refactor: supabase 관련 파일 폴더 이동
